+++ src/redteam/extraction/util.ts
